pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mr_yolo_replicate"
version = "1.0.0"
description = "MySQL database query and management tools"
requires-python = ">=3.7"
dependencies = ["Pillow>=^10.2.0"]
